# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "description": "this is paro api.",
    "version": "1.0.0",
    "title": "Swagger parto"
  },
  "servers": [
    {
      "url": "https://api.parto.app"
    },
    {
      "url": "http://localhost:2216"
    },
    {
      "url": "https://dev.parto.app"
    }
  ],
...
  "paths": {
...
    "/pregnancy/getPregnancyData/{userId}/{lang}": {
      "get": {
        "tags": [
          "pregnancy"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "description": "user id"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "lang",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "an authorization header"
                }
              }
            }
          }
        ],
        "summary": "get user pregnancy info",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "description": "user profile information",
                      "properties": {
                        "due_date": {
                          "type": "string"
                        },
                        "abortion": {
                          "type": "integer"
                        },
                        "conception_date": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "user information not found"
          }
        }
      }
    },
    "/pregnancy/savePregnancyData/{userId}/{lang}": {
      "post": {
        "tags": [
          "pregnancy"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "description": "user id"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "lang",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "an authorization header"
                }
              }
            }
          }
        ],
        "summary": "save user pregnancy info",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dueDate": {
                    "type": "string"
                  },
                  "abortion": {
                    "type": "integer"
                  },
                  "conceptionDate": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "invalid information"
          },
          "404": {
            "description": "information not found"
          },
          "409": {
            "description": "contradiction between information"
          }
        }
      }
    },
    "/pregnancy/setPregnancyEnd/{userId}/{dueDate}/{lang}": {
      "post": {
        "tags": [
          "pregnancy"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "description": "user id"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "lang",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "dueDate",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "an authorization header"
                }
              }
            }
          }
        ],
        "summary": "save user pregnancy info",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "invalid information"
          },
          "404": {
            "description": "information not found"
          },
          "409": {
            "description": "contradiction between information"
          }
        }
      }
    },
    "/cycle/getLastPeriodDate/{userId}/{lang}": {
      "get": {
        "tags": [
          "cycle"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "description": "user id"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "lang",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "an authorization header"
                }
              }
            }
          }
        ],
        "summary": "get user pregnancy info",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "description": "user profile information",
                      "properties": {
                        "last_period_date": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "user information not found"
          }
        }
      }
    },
    "/cycle/getUserAllPeriodDays/{userId}/{lang}": {
      "get": {
        "tags": [
          "cycle"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "description": "user id"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "lang",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "an authorization header"
                }
              }
            }
          }
        ],
        "summary": "get all bleeding day",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "date": [
                        "2020-07-09",
                        "2020-07-10",
                        "2020-07-11",
                        "2020-07-12"
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "user information not found"
          }
        }
      }
    },
    "/cycle/editLastPeriodDate/{userId}/{lastPeriodDate}/{lang}": {
      "put": {
        "tags": [
          "cycle"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "description": "user id"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "lang",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "lastPeriodDate",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "last Period Date"
                }
              }
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "an authorization header"
                }
              }
            }
          }
        ],
        "summary": "edit user peofile information to user",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "invalid information"
          }
        }
      }
    },
    "/cycle/setBleedingDays/{userId}/{lang}": {
      "put": {
        "tags": [
          "cycle"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "description": "user id"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "lang",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "an authorization header"
                }
              }
            }
          }
        ],
        "summary": "edit user peofile information to user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "deleteDate": [
                    "2020-07-09",
                    "2020-07-10"
                  ],
                  "addDate": [
                    "2020-07-11",
                    "2020-07-12"
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "invalid information entery"
          }
        }
      }
    }
  }
}